diffc and advecc kappa work with minor adjustments

+++ src/config.f90
@@ -14,7 +14,7 @@ module config
 
    integer :: iexpnr = -1                    ! Default value indicates it must be specified
    integer :: runtime = 25                   !< seconds
-   integer :: dtmax = 1                      !< seconds
+   real :: dtmax = 1.                      !< seconds
    integer :: integration_scheme = 1         !< 1. euler, 2.heun, 3.rk3, 4. rk4, (only 1 supported so far)
    integer :: field_load_chunk_size = 100    ! Sensible default
    character(len=256) :: field_dump_path = ''  ! Must be specified
